requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
openai==1.10.0
pydantic==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.9
gunicorn==20.1.0
pytest==7.4.4
